Extract modules expr,select,where_clause from translate.rs
